my solution to sparse array
There is a collection of input strings and a collection of query strings. For each query string, determine how many times it occurs in the list of input strings.
Complete the function matchingStrings in the editor below. The function must return an array of integers representing the frequency of occurrence of each query string in strings.

matchingStrings has the following parameters:

    strings - an array of strings to search
    queries - an array of query strings