solution to  triangle quest 2
You are given a positive integer .N
Your task is to print a palindromic triangle of sizeN
For example, a palindromic triangle of size 5
is:

1
121
12321
1234321
123454321
You can't take more than two lines. The first line (a for-statement) is already written for you.
You have to complete the code using exactly one print statement.